feat(upload-service): add worker container for background processing, merge upload and history pages, apply bycoders palette
