+++ frontend-service/src/pages/Upload.tsx
@@ -1,6 +1,5 @@
-import { useState, useEffect, useRef, useCallback } from 'react';
-import type { FC, DragEvent } from 'react';
-import { Link } from 'react-router-dom';
+import { useState, useEffect, useRef, useCallback, useMemo } from 'react';
+import type { FC, DragEvent, ChangeEvent } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
   faCloudArrowUp,
@@ -10,6 +9,13 @@ import {
   faSpinner,
   faHourglass,
   faBoxOpen,
+  faMagnifyingGlass,
+  faChevronUp,
+  faChevronDown,
+  faChevronLeft,
+  faChevronRight,
+  faAnglesLeft,
+  faAnglesRight,
 } from '@fortawesome/free-solid-svg-icons';
 import Layout from '../components/Layout';
 import { uploadService } from '../services/uploadService';
@@ -39,11 +45,38 @@ const STATUS_ICONS: Record<UploadResponse['status'], typeof faHourglass> = {
   failed: faCircleXmark,
 };
 
+// Status filter options shown in the dropdown
+const STATUS_OPTIONS: { value: string; label: string }[] = [
+  { value: '', label: 'Todos' },
+  { value: 'pending', label: 'Pendente' },
+  { value: 'processing', label: 'Processando' },
+  { value: 'completed', label: 'Concluído' },
+  { value: 'failed', label: 'Falhou' },
+];
+
+// Available page size choices for the table toolbar
+const PAGE_SIZE_OPTIONS = [5, 10, 20, 50];
+
 // Auto-refresh interval in milliseconds
 const POLL_INTERVAL_MS = 10_000;
 
-// Number of recent uploads shown in the mini-table
-const RECENT_PAGE_SIZE = 5;
+// Columns that support sorting
+type SortKey = 'original_filename' | 'status' | 'total_transactions' | 'created_at';
+type SortDir = 'asc' | 'desc';
+
+// Format bytes into human-readable string
+function formatBytes(bytes: number): string {
+  if (bytes < 1024) return `${bytes} B`;
+  if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`;
+  return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
+}
+
+// Format ISO date string to DD/MM/YYYY HH:MM
+function formatDate(iso: string): string {
+  const d = new Date(iso);
+  const pad = (n: number) => String(n).padStart(2, '0');
+  return `${pad(d.getDate())}/${pad(d.getMonth() + 1)}/${d.getFullYear()} ${pad(d.getHours())}:${pad(d.getMinutes())}`;
+}
 
 interface StatusBadgeProps {
   status: UploadResponse['status'];
@@ -62,24 +95,36 @@ const StatusBadge: FC<StatusBadgeProps> = ({ status }) => (
   </span>
 );
 
-// Format bytes into human-readable string
-function formatBytes(bytes: number): string {
-  if (bytes < 1024) return `${bytes} B`;
-  if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`;
-  return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
+interface SortIconProps {
+  columnKey: SortKey;
+  sortKey: SortKey;
+  sortDir: SortDir;
 }
 
-// Format ISO date string to DD/MM/YYYY HH:MM
-function formatDate(iso: string): string {
-  const d = new Date(iso);
-  const pad = (n: number) => String(n).padStart(2, '0');
-  return `${pad(d.getDate())}/${pad(d.getMonth() + 1)}/${d.getFullYear()} ${pad(d.getHours())}:${pad(d.getMinutes())}`;
-}
+// Renders the sort direction arrow for a column header
+const SortIcon: FC<SortIconProps> = ({ columnKey, sortKey, sortDir }) => {
+  if (columnKey !== sortKey) {
+    return (
+      <FontAwesomeIcon
+        icon={faChevronDown}
+        className="ml-1 text-[#3a3a3a] text-[10px]"
+        aria-hidden="true"
+      />
+    );
+  }
+  return (
+    <FontAwesomeIcon
+      icon={sortDir === 'asc' ? faChevronUp : faChevronDown}
+      className="ml-1 text-[#4FFA7B] text-[10px]"
+      aria-hidden="true"
+    />
+  );
+};
 
-// Skeleton row shown while loading the recent uploads table
+// Skeleton row shown during loading
 const SkeletonRow: FC = () => (
   <tr className="border-b border-white/5 animate-pulse">
-    {[70, 80, 50, 100].map((w, i) => (
+    {[60, 80, 50, 120, 100].map((w, i) => (
       <td key={i} className="px-4 py-3">
         <div className="h-3 bg-[#2a2a2a] rounded" style={{ width: `${w}%` }} />
       </td>
@@ -88,44 +133,145 @@ const SkeletonRow: FC = () => (
 );
 
 const Upload: FC = () => {
+  // Upload zone state
   const [selectedFile, setSelectedFile] = useState<File | null>(null);
   const [isDragOver, setIsDragOver] = useState(false);
   const [uploading, setUploading] = useState(false);
   const [successMessage, setSuccessMessage] = useState<string | null>(null);
   const [errorMessage, setErrorMessage] = useState<string | null>(null);
-  const [recentUploads, setRecentUploads] = useState<UploadResponse[]>([]);
-  const [totalCount, setTotalCount] = useState(0);
-  const [loadingRecent, setLoadingRecent] = useState(true);
 
   const fileInputRef = useRef<HTMLInputElement>(null);
 
-  // Fetch the most recent uploads for the mini-table
-  const loadRecentUploads = useCallback(async () => {
+  // DataTable state
+  const [uploads, setUploads] = useState<UploadResponse[]>([]);
+  const [loading, setLoading] = useState(true);
+  const [statusFilter, setStatusFilter] = useState('');
+  const [searchQuery, setSearchQuery] = useState('');
+  const [sortKey, setSortKey] = useState<SortKey>('created_at');
+  const [sortDir, setSortDir] = useState<SortDir>('desc');
+  const [currentPage, setCurrentPage] = useState(1);
+  const [pageSize, setPageSize] = useState(PAGE_SIZE_OPTIONS[1]);
+
+  // Fetch all uploads for the selected status from the API
+  // We use a large page_size (1000) so client-side search and pagination work on the full dataset
+  const loadUploads = useCallback(async (status: string) => {
     try {
-      const data = await uploadService.list(1, RECENT_PAGE_SIZE);
-      setRecentUploads(data.results);
-      setTotalCount(data.count);
+      setLoading(true);
+      const data = await uploadService.list(1, 1000, status || undefined);
+      setUploads(data.results);
     } catch {
-      // Silent failure — recent uploads section is non-critical
+      setUploads([]);
     } finally {
-      setLoadingRecent(false);
+      setLoading(false);
     }
   }, []);
 
-  // Initial fetch
+  // Initial fetch and refetch on status filter change
   useEffect(() => {
-    loadRecentUploads();
-  }, [loadRecentUploads]);
+    loadUploads(statusFilter);
+    setCurrentPage(1);
+    setSearchQuery('');
+  }, [statusFilter, loadUploads]);
 
   // Auto-refresh polling to catch status changes (pending -> processing -> completed)
   useEffect(() => {
     const interval = setInterval(() => {
-      loadRecentUploads();
+      loadUploads(statusFilter);
     }, POLL_INTERVAL_MS);
 
     return () => clearInterval(interval);
-  }, [loadRecentUploads]);
+  }, [loadUploads, statusFilter]);
 
+  // Apply text search filter client-side
+  const filteredUploads = useMemo(() => {
+    const q = searchQuery.trim().toLowerCase();
+    if (!q) return uploads;
+    return uploads.filter(
+      (u) =>
+        u.original_filename.toLowerCase().includes(q) ||
+        STATUS_LABELS[u.status].toLowerCase().includes(q),
+    );
+  }, [uploads, searchQuery]);
+
+  // Apply client-side sorting
+  const sortedUploads = useMemo(() => {
+    return [...filteredUploads].sort((a, b) => {
+      let valA: string | number = a[sortKey] ?? '';
+      let valB: string | number = b[sortKey] ?? '';
+
+      if (sortKey === 'total_transactions') {
+        valA = Number(valA);
+        valB = Number(valB);
+        return sortDir === 'asc' ? (valA as number) - (valB as number) : (valB as number) - (valA as number);
+      }
+
+      // String comparison for other columns
+      valA = String(valA).toLowerCase();
+      valB = String(valB).toLowerCase();
+      if (valA < valB) return sortDir === 'asc' ? -1 : 1;
+      if (valA > valB) return sortDir === 'asc' ? 1 : -1;
+      return 0;
+    });
+  }, [filteredUploads, sortKey, sortDir]);
+
+  // Pagination calculations derived from sorted data
+  const totalFiltered = sortedUploads.length;
+  const totalPages = Math.max(1, Math.ceil(totalFiltered / pageSize));
+  const safePage = Math.min(currentPage, totalPages);
+  const startIndex = (safePage - 1) * pageSize;
+  const endIndex = Math.min(startIndex + pageSize, totalFiltered);
+  const pageRows = sortedUploads.slice(startIndex, endIndex);
+
+  // Compute visible page buttons (max 5 around current page)
+  const pageNumbers = useMemo(() => {
+    const range: number[] = [];
+    const delta = 2;
+    const left = Math.max(1, safePage - delta);
+    const right = Math.min(totalPages, safePage + delta);
+    for (let i = left; i <= right; i++) range.push(i);
+    return range;
+  }, [safePage, totalPages]);
+
+  const handleSortColumn = (key: SortKey) => {
+    if (sortKey === key) {
+      setSortDir((d) => (d === 'asc' ? 'desc' : 'asc'));
+    } else {
+      setSortKey(key);
+      setSortDir('asc');
+    }
+    setCurrentPage(1);
+  };
+
+  const handleSearchChange = (e: ChangeEvent<HTMLInputElement>) => {
+    setSearchQuery(e.target.value);
+    setCurrentPage(1);
+  };
+
+  const handleStatusChange = (e: ChangeEvent<HTMLSelectElement>) => {
+    setStatusFilter(e.target.value);
+  };
+
+  const handlePageSizeChange = (e: ChangeEvent<HTMLSelectElement>) => {
+    setPageSize(Number(e.target.value));
+    setCurrentPage(1);
+  };
+
+  // Shared header cell renderer for sortable columns
+  const SortableHeader: FC<{ label: string; colKey: SortKey }> = ({ label, colKey }) => (
+    <th
+      scope="col"
+      className="px-4 py-3 text-left text-xs font-semibold uppercase tracking-wider text-[#898989] cursor-pointer hover:text-white select-none transition-colors"
+      onClick={() => handleSortColumn(colKey)}
+      aria-sort={sortKey === colKey ? (sortDir === 'asc' ? 'ascending' : 'descending') : 'none'}
+    >
+      <span className="inline-flex items-center gap-1">
+        {label}
+        <SortIcon columnKey={colKey} sortKey={sortKey} sortDir={sortDir} />
+      </span>
+    </th>
+  );
+
+  // Upload zone handlers
   const handleFileChange = (file: File | null) => {
     setSelectedFile(file);
     setSuccessMessage(null);
@@ -171,7 +317,8 @@ const Upload: FC = () => {
       setSelectedFile(null);
       // Reset the hidden file input so the same file can be selected again
       if (fileInputRef.current) fileInputRef.current.value = '';
-      await loadRecentUploads();
+      // Refresh the DataTable to show the new upload
+      await loadUploads(statusFilter);
     } catch {
       setErrorMessage('Ocorreu um erro ao enviar o arquivo. Tente novamente.');
     } finally {
@@ -190,7 +337,7 @@ const Upload: FC = () => {
           </p>
         </div>
 
-        {/* Drop zone card — unchanged */}
+        {/* Drop zone card */}
         <div className="bg-[#1e1e1e] border border-[#2a2a2a] rounded-xl p-5">
           {/* Hidden file input — accepts .txt and .cnab */}
           <input
@@ -274,124 +421,234 @@ const Upload: FC = () => {
           )}
         </div>
 
-        {/* Recent uploads mini-table */}
-        <div>
-          {/* Section header with count badge and link to full history */}
-          <div className="flex items-center justify-between mb-3">
-            <div className="flex items-center gap-2">
-              <h3 className="text-[#FFFFFF] text-base font-semibold">Uploads Recentes</h3>
-              {!loadingRecent && totalCount > 0 && (
-                <span className="bg-[#2a2a2a] text-[#D8D8D8] text-xs font-medium px-2 py-0.5 rounded-full">
-                  {totalCount}
-                </span>
-              )}
+        {/* History section header with status filter */}
+        <div className="flex items-center justify-between">
+          <h3 className="text-[#FFFFFF] text-base font-semibold">Histórico de Uploads</h3>
+          <div className="flex items-center gap-2">
+            <label htmlFor="status-filter" className="text-[#D8D8D8] text-sm shrink-0">
+              Status:
+            </label>
+            <select
+              id="status-filter"
+              value={statusFilter}
+              onChange={handleStatusChange}
+              className="bg-[#2a2a2a] border border-[#3a3a3a] text-[#FFFFFF] text-sm rounded-md px-3 py-1.5 focus:outline-none focus:ring-1 focus:ring-[#4FFA7B] cursor-pointer"
+              aria-label="Filtrar por status"
+            >
+              {STATUS_OPTIONS.map((opt) => (
+                <option key={opt.value} value={opt.value}>
+                  {opt.label}
+                </option>
+              ))}
+            </select>
+          </div>
+        </div>
+
+        {/* DataTable container */}
+        <div className="bg-[#1e1e1e] border border-white/5 rounded-xl overflow-hidden">
+
+          {/* Table toolbar: search on left, page size on right */}
+          <div className="flex flex-col sm:flex-row sm:items-center sm:justify-between gap-3 px-4 py-3 border-b border-white/5">
+            {/* Search input */}
+            <div className="relative flex-1 max-w-xs">
+              <span className="absolute inset-y-0 left-3 flex items-center pointer-events-none text-[#898989]">
+                <FontAwesomeIcon icon={faMagnifyingGlass} className="text-xs" aria-hidden="true" />
+              </span>
+              <input
+                type="search"
+                placeholder="Buscar por arquivo ou status..."
+                value={searchQuery}
+                onChange={handleSearchChange}
+                className="w-full bg-[#171616] border border-white/10 rounded-lg pl-8 pr-3 py-2 text-sm text-white placeholder-[#898989] focus:outline-none focus:ring-1 focus:ring-[#4FFA7B]"
+                aria-label="Buscar uploads"
+              />
+            </div>
+
+            {/* Page size selector */}
+            <div className="flex items-center gap-2 shrink-0">
+              <span className="text-[#898989] text-xs">Exibir</span>
+              <select
+                value={pageSize}
+                onChange={handlePageSizeChange}
+                className="bg-[#171616] border border-white/10 rounded-lg px-2 py-1.5 text-xs text-white focus:outline-none focus:ring-1 focus:ring-[#4FFA7B] cursor-pointer"
+                aria-label="Itens por página"
+              >
+                {PAGE_SIZE_OPTIONS.map((size) => (
+                  <option key={size} value={size}>
+                    {size}
+                  </option>
+                ))}
+              </select>
+              <span className="text-[#898989] text-xs">por página</span>
             </div>
           </div>
 
-          {/* DataTable-style container */}
-          <div className="bg-[#1e1e1e] border border-white/5 rounded-xl overflow-hidden">
-            <div className="overflow-x-auto">
-              <table className="w-full" aria-label="Uploads recentes">
-                <thead>
-                  <tr className="border-b border-white/5 bg-[#171616]/50">
-                    <th
-                      scope="col"
-                      className="px-4 py-3 text-left text-xs font-semibold uppercase tracking-wider text-[#898989] select-none"
-                    >
-                      Arquivo
-                    </th>
-                    <th
-                      scope="col"
-                      className="px-4 py-3 text-left text-xs font-semibold uppercase tracking-wider text-[#898989] select-none"
-                    >
-                      Status
-                    </th>
-                    <th
-                      scope="col"
-                      className="px-4 py-3 text-left text-xs font-semibold uppercase tracking-wider text-[#898989] select-none"
-                    >
-                      Transações
-                    </th>
-                    <th
-                      scope="col"
-                      className="px-4 py-3 text-left text-xs font-semibold uppercase tracking-wider text-[#898989] select-none"
-                    >
-                      Data
-                    </th>
+          {/* Table */}
+          <div className="overflow-x-auto">
+            <table className="w-full" aria-label="Histórico de uploads">
+              <thead>
+                <tr className="border-b border-white/5 bg-[#171616]/50">
+                  <SortableHeader label="Arquivo" colKey="original_filename" />
+                  <SortableHeader label="Status" colKey="status" />
+                  <SortableHeader label="Transações" colKey="total_transactions" />
+                  {/* Error column is not sortable */}
+                  <th
+                    scope="col"
+                    className="px-4 py-3 text-left text-xs font-semibold uppercase tracking-wider text-[#898989] select-none"
+                  >
+                    Erro
+                  </th>
+                  <SortableHeader label="Data de Upload" colKey="created_at" />
+                </tr>
+              </thead>
+              <tbody>
+                {loading ? (
+                  // Show skeleton rows while fetching
+                  Array.from({ length: 5 }).map((_, i) => <SkeletonRow key={i} />)
+                ) : pageRows.length === 0 ? (
+                  // Empty state spanning all columns
+                  <tr>
+                    <td colSpan={5} className="px-4 py-16 text-center">
+                      <div className="flex flex-col items-center gap-3">
+                        <FontAwesomeIcon
+                          icon={faBoxOpen}
+                          className="text-[#D8D8D8] text-4xl opacity-30"
+                          aria-hidden="true"
+                        />
+                        <p className="text-[#D8D8D8] font-medium">
+                          Nenhum upload encontrado
+                        </p>
+                        <p className="text-[#898989] text-sm">
+                          {searchQuery
+                            ? 'Nenhum resultado para a busca realizada.'
+                            : statusFilter
+                            ? 'Nenhum upload com o status selecionado.'
+                            : 'Importe um arquivo CNAB para começar.'}
+                        </p>
+                      </div>
+                    </td>
                   </tr>
-                </thead>
-                <tbody>
-                  {loadingRecent ? (
-                    // Skeleton rows while fetching
-                    Array.from({ length: RECENT_PAGE_SIZE }).map((_, i) => (
-                      <SkeletonRow key={i} />
-                    ))
-                  ) : recentUploads.length === 0 ? (
-                    // Empty state
-                    <tr>
-                      <td colSpan={4} className="px-4 py-12 text-center">
-                        <div className="flex flex-col items-center gap-3">
-                          <FontAwesomeIcon
-                            icon={faBoxOpen}
-                            className="text-[#D8D8D8] text-3xl opacity-30"
-                            aria-hidden="true"
-                          />
-                          <p className="text-[#D8D8D8] font-medium text-sm">
-                            Nenhum upload encontrado
-                          </p>
-                          <p className="text-[#898989] text-xs">
-                            Importe um arquivo CNAB para começar.
-                          </p>
-                        </div>
+                ) : (
+                  pageRows.map((upload) => (
+                    <tr
+                      key={upload.id}
+                      className="border-b border-white/5 last:border-0 hover:bg-white/[0.02] transition-colors"
+                    >
+                      {/* Filename */}
+                      <td
+                        className="px-4 py-3 text-sm text-[#FFFFFF] font-medium max-w-[200px] truncate"
+                        title={upload.original_filename}
+                      >
+                        {upload.original_filename}
+                      </td>
+
+                      {/* Status badge */}
+                      <td className="px-4 py-3">
+                        <StatusBadge status={upload.status} />
+                      </td>
+
+                      {/* Transaction count */}
+                      <td className="px-4 py-3 text-sm text-[#D8D8D8] tabular-nums">
+                        {upload.total_transactions}
+                      </td>
+
+                      {/* Error message — truncated with title tooltip */}
+                      <td
+                        className="px-4 py-3 text-sm text-[#FF4444] max-w-[200px] truncate"
+                        title={upload.error_message ?? undefined}
+                      >
+                        {upload.error_message ? (
+                          <span className="truncate block">{upload.error_message}</span>
+                        ) : (
+                          <span className="text-[#898989]">—</span>
+                        )}
+                      </td>
+
+                      {/* Upload date */}
+                      <td className="px-4 py-3 text-sm text-[#D8D8D8] tabular-nums whitespace-nowrap">
+                        {formatDate(upload.created_at)}
                       </td>
                     </tr>
-                  ) : (
-                    recentUploads.map((upload) => (
-                      <tr
-                        key={upload.id}
-                        className="border-b border-white/5 last:border-0 hover:bg-white/[0.02] transition-colors"
-                      >
-                        {/* Filename */}
-                        <td
-                          className="px-4 py-3 text-sm text-[#FFFFFF] font-medium max-w-[180px] truncate"
-                          title={upload.original_filename}
-                        >
-                          {upload.original_filename}
-                        </td>
-
-                        {/* Status badge */}
-                        <td className="px-4 py-3">
-                          <StatusBadge status={upload.status} />
-                        </td>
-
-                        {/* Transaction count */}
-                        <td className="px-4 py-3 text-sm text-[#D8D8D8] tabular-nums">
-                          {upload.total_transactions}
-                        </td>
-
-                        {/* Upload date */}
-                        <td className="px-4 py-3 text-sm text-[#D8D8D8] tabular-nums whitespace-nowrap">
-                          {formatDate(upload.created_at)}
-                        </td>
-                      </tr>
-                    ))
-                  )}
-                </tbody>
-              </table>
-            </div>
-
-            {/* Footer with link to full history */}
-            {!loadingRecent && (
-              <div className="px-4 py-3 border-t border-white/5 flex justify-end">
-                <Link
-                  to="/history"
-                  className="text-[#4FFA7B] hover:text-[#02BE3B] text-sm font-medium transition-colors"
-                >
-                  Ver histórico completo →
-                </Link>
-              </div>
-            )}
+                  ))
+                )}
+              </tbody>
+            </table>
           </div>
+
+          {/* Table footer: record count on left, pagination on right */}
+          {!loading && (
+            <div className="flex flex-col sm:flex-row sm:items-center sm:justify-between gap-3 px-4 py-3 border-t border-white/5">
+              {/* Record range summary */}
+              <p className="text-xs text-[#898989]">
+                {totalFiltered === 0
+                  ? 'Nenhum registro encontrado'
+                  : `Exibindo ${startIndex + 1}–${endIndex} de ${totalFiltered} registro${totalFiltered !== 1 ? 's' : ''}`}
+              </p>
+
+              {/* Page navigation buttons */}
+              {totalPages > 1 && (
+                <div className="flex items-center gap-1" role="navigation" aria-label="Paginação">
+                  {/* First page */}
+                  <button
+                    onClick={() => setCurrentPage(1)}
+                    disabled={safePage === 1}
+                    className="w-8 h-8 flex items-center justify-center text-[#898989] hover:text-white hover:bg-white/5 rounded-lg transition-colors disabled:opacity-30 disabled:cursor-not-allowed"
+                    aria-label="Primeira página"
+                  >
+                    <FontAwesomeIcon icon={faAnglesLeft} className="text-xs" aria-hidden="true" />
+                  </button>
+
+                  {/* Previous page */}
+                  <button
+                    onClick={() => setCurrentPage((p) => Math.max(1, p - 1))}
+                    disabled={safePage === 1}
+                    className="w-8 h-8 flex items-center justify-center text-[#898989] hover:text-white hover:bg-white/5 rounded-lg transition-colors disabled:opacity-30 disabled:cursor-not-allowed"
+                    aria-label="Página anterior"
+                  >
+                    <FontAwesomeIcon icon={faChevronLeft} className="text-xs" aria-hidden="true" />
+                  </button>
+
+                  {/* Numbered page buttons */}
+                  {pageNumbers.map((num) => (
+                    <button
+                      key={num}
+                      onClick={() => setCurrentPage(num)}
+                      className={[
+                        'w-8 h-8 flex items-center justify-center rounded-lg text-xs transition-colors',
+                        num === safePage
+                          ? 'bg-[#02BE3B] text-white font-semibold'
+                          : 'text-[#898989] hover:text-white hover:bg-white/5',
+                      ].join(' ')}
+                      aria-label={`Página ${num}`}
+                      aria-current={num === safePage ? 'page' : undefined}
+                    >
+                      {num}
+                    </button>
+                  ))}
+
+                  {/* Next page */}
+                  <button
+                    onClick={() => setCurrentPage((p) => Math.min(totalPages, p + 1))}
+                    disabled={safePage === totalPages}
+                    className="w-8 h-8 flex items-center justify-center text-[#898989] hover:text-white hover:bg-white/5 rounded-lg transition-colors disabled:opacity-30 disabled:cursor-not-allowed"
+                    aria-label="Próxima página"
+                  >
+                    <FontAwesomeIcon icon={faChevronRight} className="text-xs" aria-hidden="true" />
+                  </button>
+
+                  {/* Last page */}
+                  <button
+                    onClick={() => setCurrentPage(totalPages)}
+                    disabled={safePage === totalPages}
+                    className="w-8 h-8 flex items-center justify-center text-[#898989] hover:text-white hover:bg-white/5 rounded-lg transition-colors disabled:opacity-30 disabled:cursor-not-allowed"
+                    aria-label="Última página"
+                  >
+                    <FontAwesomeIcon icon={faAnglesRight} className="text-xs" aria-hidden="true" />
+                  </button>
+                </div>
+              )}
+            </div>
+          )}
         </div>
       </div>
     </Layout>
